Hack out all the room generation stuff.
Preparing for yet another different world generation strategy.
