+++ src/main/java/com/emergentmud/core/command/impl/MapCommand.java
@@ -22,9 +22,9 @@ package com.emergentmud.core.command.impl;
 
 import com.emergentmud.core.command.BaseCommand;
 import com.emergentmud.core.model.Entity;
-import com.emergentmud.core.model.room.Room;
+import com.emergentmud.core.model.Room;
 import com.emergentmud.core.model.stomp.GameOutput;
-import com.emergentmud.core.repository.RoomBuilder;
+import com.emergentmud.core.repository.RoomRepository;
 import org.springframework.stereotype.Component;
 
 import javax.inject.Inject;
@@ -34,18 +34,18 @@ public class MapCommand extends BaseCommand {
     private static final int MAP_EXTENT_X = 40;
     private static final int MAP_EXTENT_Y = 20;
 
-    private RoomBuilder roomBuilder;
+    private RoomRepository roomRepository;
 
     @Inject
-    public MapCommand(RoomBuilder roomBuilder) {
-        this.roomBuilder = roomBuilder;
+    public MapCommand(RoomRepository roomRepository) {
+        this.roomRepository = roomRepository;
 
         setDescription("Shows a bird's eye view of the rooms around you.");
     }
 
     @Override
     public GameOutput execute(GameOutput output, Entity entity, String command, String[] tokens, String raw) {
-        Room center = roomBuilder.generateRoom(entity.getX(), entity.getY(), entity.getZ());
+        Room center = roomRepository.findByXAndYAndZ(entity.getX(), entity.getY(), entity.getZ());
 
         for (long y = center.getY() + MAP_EXTENT_Y, i = 0; y >= center.getY() - MAP_EXTENT_Y; y--, i++) {
             StringBuilder line = new StringBuilder();
@@ -54,11 +54,10 @@ public class MapCommand extends BaseCommand {
                 if (x == center.getX() && y == center.getY()) {
                     line.append("[cyan][]</span>");
                 } else {
-                    Room room = roomBuilder.generateRoom(x, y, center.getZ());
+                    Room room = roomRepository.findByXAndYAndZ(x, y, center.getZ());
 
                     if (room != null) {
-                        line.append(String.format("<span style='color: #%02x'>[]</span>",
-                                room.getBiome() != null ? room.getBiome().getColor() : 0xFF00FF));
+                        line.append(String.format("<span style='color: #%02x'>[]</span>", 0xFF00FF));
                     } else {
                         line.append(String.format("<span style='color: #%02x%02x%02x'>&nbsp;&nbsp;</span>", 0, 0, 0));
                     }
