+++ src/main/java/com/emergentmud/core/repository/BiomeRepository.java
@@ -20,7 +20,7 @@
 
 package com.emergentmud.core.repository;
 
-import com.emergentmud.core.model.room.Biome;
+import com.emergentmud.core.model.Biome;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
