+++ src/main/java/com/emergentmud/core/repository/loader/BiomeLoader.java
@@ -21,7 +21,7 @@
 package com.emergentmud.core.repository.loader;
 
 import com.emergentmud.core.model.WhittakerGridLocation;
-import com.emergentmud.core.model.room.Biome;
+import com.emergentmud.core.model.Biome;
 import com.emergentmud.core.repository.BiomeRepository;
 import com.emergentmud.core.repository.WhittakerGridLocationRepository;
 import org.slf4j.Logger;
