+++ src/test/java/com/emergentmud/core/command/impl/ExileCommandTest.java
@@ -24,7 +24,7 @@ import com.emergentmud.core.model.Account;
 import com.emergentmud.core.model.Capability;
 import com.emergentmud.core.model.CommandRole;
 import com.emergentmud.core.model.Entity;
-import com.emergentmud.core.model.room.Room;
+import com.emergentmud.core.model.Room;
 import com.emergentmud.core.model.stomp.GameOutput;
 import com.emergentmud.core.repository.AccountRepository;
 import com.emergentmud.core.repository.CapabilityRepository;
