+++ src/test/java/com/emergentmud/core/command/impl/LookCommandTest.java
@@ -21,11 +21,11 @@
 package com.emergentmud.core.command.impl;
 
 import com.emergentmud.core.model.Entity;
-import com.emergentmud.core.model.room.Biome;
-import com.emergentmud.core.model.room.Room;
+import com.emergentmud.core.model.Biome;
+import com.emergentmud.core.model.Room;
 import com.emergentmud.core.model.stomp.GameOutput;
 import com.emergentmud.core.repository.EntityRepository;
-import com.emergentmud.core.repository.RoomBuilder;
+import com.emergentmud.core.repository.RoomRepository;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -39,7 +39,7 @@ public class LookCommandTest {
     private EntityRepository entityRepository;
 
     @Mock
-    private RoomBuilder roomBuilder;
+    private RoomRepository roomRepository;
 
     @Mock
     private GameOutput output;
@@ -63,7 +63,7 @@ public class LookCommandTest {
     public void setUp() throws Exception {
         MockitoAnnotations.initMocks(this);
 
-        command = new LookCommand(entityRepository, roomBuilder);
+        command = new LookCommand(entityRepository, roomRepository);
     }
 
     @Test
@@ -86,10 +86,7 @@ public class LookCommandTest {
 
     @Test
     public void testLook() throws Exception {
-        when(roomBuilder.generateRoom(eq(0L), eq(0L), eq(0L))).thenReturn(room);
-        when(room.getBiome()).thenReturn(biome);
-        when(room.getElevation()).thenReturn(1);
-        when(room.getMoisture()).thenReturn(0);
+        when(roomRepository.findByXAndYAndZ(eq(0L), eq(0L), eq(0L))).thenReturn(room);
         when(biome.getName()).thenReturn("Blasted Hellscape");
