+++ src/test/java/com/emergentmud/core/command/impl/TeleportCommandTest.java
@@ -21,7 +21,7 @@
 package com.emergentmud.core.command.impl;
 
 import com.emergentmud.core.model.Entity;
-import com.emergentmud.core.model.room.Room;
+import com.emergentmud.core.model.Room;
 import com.emergentmud.core.model.stomp.GameOutput;
 import com.emergentmud.core.repository.WorldManager;
 import com.emergentmud.core.service.EntityService;
