+++ src/test/java/com/emergentmud/core/event/StompDisconnectListenerTest.java
@@ -21,7 +21,7 @@
 package com.emergentmud.core.event;
 
 import com.emergentmud.core.model.Entity;
-import com.emergentmud.core.model.room.Room;
+import com.emergentmud.core.model.Room;
 import com.emergentmud.core.model.stomp.GameOutput;
 import com.emergentmud.core.repository.EntityRepository;
 import com.emergentmud.core.repository.WorldManager;
