--- src/test/java/com/emergentmud/core/model/room/BiomeTest.java
@@ -1,57 +1,0 @@
-/*
- * EmergentMUD - A modern MUD with a procedurally generated world.
- * Copyright (C) 2016-2017 Peter Keeler
- *
- * This file is part of EmergentMUD.
- *
- * EmergentMUD is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * EmergentMUD is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package com.emergentmud.core.model.room;
-
-import org.junit.Before;
-import org.junit.Test;
-
-import java.util.UUID;
-
-import static org.junit.Assert.*;
-
-public class BiomeTest {
-    private Biome biome;
-
-    @Before
-    public void setUp() throws Exception {
-        biome = new Biome("Biome", 0x112233);
-    }
-
-    @Test
-    public void testId() throws Exception {
-        String guid = UUID.randomUUID().toString();
-
-        biome.setId("foo");
-        assertEquals("foo", biome.getId());
-        biome.setId(guid);
-        assertEquals(guid, biome.getId());
-    }
-
-    @Test
-    public void testName() throws Exception {
-        assertEquals("Biome", biome.getName());
-    }
-
-    @Test
-    public void testColor() throws Exception {
-        assertEquals(0x112233, (long)biome.getColor());
-    }
-}
