+++ src/test/java/com/emergentmud/core/repository/loader/BiomeLoaderTest.java
@@ -20,7 +20,7 @@
 
 package com.emergentmud.core.repository.loader;
 
-import com.emergentmud.core.model.room.Biome;
+import com.emergentmud.core.model.Biome;
 import com.emergentmud.core.repository.BiomeRepository;
 import com.emergentmud.core.repository.RoomRepository;
 import com.emergentmud.core.repository.WhittakerGridLocationRepository;
